% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_functional_response.R
\name{bin_vec}
\alias{bin_vec}
\title{Bin a vector.}
\usage{
bin_vec(x, binsize)
}
\arguments{
\item{x}{The numeric vector.}

\item{binsize}{The binsize}
}
\value{
Vector binned.
}
\description{
Bin a vector.
}
