% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_functional_response.R
\name{do_read_data}
\alias{do_read_data}
\title{Read in simulation data.}
\usage{
do_read_data(
  data_folder,
  which_gen = seq(991, 998, 1),
  n_time = 400,
  layers = c("items", "foragers", "klepts", "klepts_intake", "foragers_intake")
)
}
\arguments{
\item{data_folder}{Folder to read from.}

\item{which_gen}{Generations to read. Numeric sequence.}

\item{n_time}{Number of timesteps. Single numeric.}

\item{layers}{Which layers to add.}
}
\value{
A list of data.tables each with the cell specific,
per-timestep value of a single variable in each of the generations
queried.
}
\description{
Read in simulation data.
}
