% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_functional_response.R
\name{get_functional_response}
\alias{get_functional_response}
\title{Get generalised functional response data.}
\usage{
get_functional_response(
  response = c("pc_intake_forager", "pc_intake_klepts", "pc_intake_total"),
  drivers = c("klepts", "foragers", "total_agents", "items"),
  round_value = 0.005,
  data_folder,
  which_gen = seq(991, 998, 1),
  n_time = 200,
  layers = c("items", "foragers", "klepts", "klepts_intake", "foragers_intake")
)
}
\arguments{
\item{data_folder}{Which data folder to summarise. Data folders should
contain the results of ONE replicate of ONE parameter combination.}

\item{which_gen}{Which generations to look for. Defaults to 991 -- 998.}

\item{n_time}{The timesteps per generation.}

\item{layers}{Which layers to read in. Defaults to all layers.}
}
\value{
A data.table with the functional response over cells grouped
by items and total number of agents.
}
\description{
Get generalised functional response data.
}
