% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_weight_evo.R
\name{get_generation_data}
\alias{get_generation_data}
\title{Get generation data.}
\usage{
get_generation_data(filepath, which_gen = 991)
}
\arguments{
\item{filepath}{Simulation output path.}

\item{which_gen}{Which generation to get.}
}
\value{
A \code{generation} object.
}
\description{
Get generation data.
}
