% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_functional_response.R
\name{get_intake_variance}
\alias{get_intake_variance}
\title{Get intake variance.}
\usage{
get_intake_variance(
  response = c("pc_intake_forager", "pc_intake_klepts", "pc_intake_total"),
  data_folder,
  which_gen = seq(991, 998, 1),
  n_time = 400
)
}
\arguments{
\item{response}{Which per capita intake to get variance for.}

\item{data_folder}{The data folder.}

\item{which_gen}{Which generations.}

\item{n_time}{How many timesteps per gen. Kleptomove runs with 400.}
}
\value{

}
\description{
Get intake variance.
}
