% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_potential_intake.R
\name{get_potential_intake}
\alias{get_potential_intake}
\title{Get potential intake rate from item count.}
\usage{
get_potential_intake(p_detection, n_items)
}
\arguments{
\item{p_detection}{Prob. detecting a single item.}

\item{n_items}{Items on the cell.}
}
\value{
A potential intake rate.
}
\description{
Get potential intake rate from item count.
}
