% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_syndrome.R
\name{get_pref_handler_by_strat}
\alias{get_pref_handler_by_strat}
\title{Get handler preference by strategy.}
\usage{
get_pref_handler_by_strat(
  data_folder,
  generations,
  weight_comp_strat,
  weight_of_interest,
  handler_pref_by_strategy = FALSE
)
}
\arguments{
\item{data_folder}{The data folder, should refer to an obligate simulation.}

\item{generations}{The generations to extact, recommended 1 - 100.}

\item{weight_comp_strat}{Which weight (mostly wt_5) refers to the
inherited preference that controls the competition strategy in scenario 2.}

\item{weight_of_interest}{Which weight are we interested in. A named vector
of weight numbers.}
}
\value{
A data.table of counts of handler preference across strategy.
}
\description{
Get handler preference by strategy.
}
