% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_scaled_wts.R
\name{get_scaled_move_prefs}
\alias{get_scaled_move_prefs}
\title{Get generation data.}
\usage{
get_scaled_move_prefs(filepath, gens, weights = c(2, 3, 4), n_agents = 100)
}
\arguments{
\item{filepath}{Simulation output path.}

\item{gens}{Which generations to get.}

\item{n_agents}{How many agents to get. Max is 10,000.}
}
\value{
A \code{generation} object.
}
\description{
Get generation data.
}
