% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_strategy_gen.r
\name{get_strategy_gen}
\alias{get_strategy_gen}
\title{Get strategy counts over generations.}
\usage{
get_strategy_gen(filepath, n_agents = 10000, n_time = 400)
}
\arguments{
\item{filepath}{The filepath to data. Must contain a \code{sourceMe.R}.}
}
\value{
A data.table of strategy counts over generations.
}
\description{
Get strategy counts over generations.
}
