% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_weight_evo.R
\name{get_weights_prop}
\alias{get_weights_prop}
\title{Get weight data from \code{generation} output.}
\usage{
get_weights_prop(gen_data, digits = 2)
}
\arguments{
\item{gen_data}{A list object produced by \code{generation()}.}

\item{digits}{Rounding digits when scaling preferences.}
}
\value{
Returns a \code{data.table}
with columns \code{weight_id}, \code{weight_value} (-1 -- +1), and
\code{weight_prop}.
}
\description{
A function to return scaled weight proportion data from weight values
returned by the \code{generation} function. Returns a \code{data.table}
with columns \code{weight_id}, \code{weight_value} (-1 -- +1), and
\code{weight_prop}.
}
