% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_weight_evo.R
\name{prepare_extractor}
\alias{prepare_extractor}
\title{Prepare the Kleptomove extractor.}
\usage{
prepare_extractor(filepath)
}
\arguments{
\item{filepath}{The simulation filepath.}
}
\value{
An environment containing extractor functions.
}
\description{
Prepare the Kleptomove extractor.
}
