% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read_land.R
\name{read_landscape}
\alias{read_landscape}
\title{Read landscape data from PNG output.}
\usage{
read_landscape(
  landscape_file,
  layer,
  crop_dim,
  type = c("items", "gradient"),
  max_K = 5
)
}
\arguments{
\item{landscape_file}{Path to the landscape file.}

\item{layer}{Which layer to read.}

\item{crop_dim}{How many rows and columns to return.}

\item{type}{What to return, either "items" for the item counts, or "gradient"
for the gradient in item counts.}

\item{max_K}{Carrying capacity of a cell.}
}
\value{
A data.table of X, Y and value
}
\description{
Read landscape data from PNG output.
}
