% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_theme.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom theme with grey background}
\usage{
theme_custom(grid = FALSE)
}
\arguments{
\item{grid}{Show panel grid major?}
}
\value{
None
}
\description{
Custom theme with grey background
}
