% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read_land.R
\name{get_layer_variance}
\alias{get_layer_variance}
\title{Get layer variance.}
\usage{
get_layer_variance(
  landscape_file,
  layer,
  crop_dim,
  type = c("items", "gradient"),
  max_K = 5
)
}
\arguments{
\item{landscape_file}{Path to the landscape file.}

\item{layer}{Which layer to read.}

\item{crop_dim}{How many rows and columns to return.}

\item{max_K}{Carrying capacity of a cell.}
}
\value{
A single layer variance.
}
\description{
Get layer variance.
}
