% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_syndrome.R
\name{get_pref_handler_by_strat}
\alias{get_pref_handler_by_strat}
\title{Get handler preference by strategy.}
\usage{
get_pref_handler_by_strat(
  data_folder,
  generations,
  weight_klept_bias,
  weight_of_interest
)
}
\arguments{
\item{data_folder}{The data folder, should refer to an obligate simulation.}

\item{generations}{The generations to extact, recommended 1 - 100.}

\item{weight_klept_bias}{Which weight (mostly 5 or 7) refers to the kleptoparasite
strategy, this is the bias in obligate simulations.}

\item{weight_of_interest}{Which weight are we interested in. A named vector
of weight numbers.}
}
\value{
A data.table of counts of handler preference across strategy.
}
\description{
Get handler preference by strategy.
}
