% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_weight_evo.R
\name{get_sim_weight_evol}
\alias{get_sim_weight_evol}
\title{Get weight evolution data for a simulation replicate.}
\usage{
get_sim_weight_evol(data_folder, weight_names = c(), ...)
}
\arguments{
\item{data_folder}{Where the \code{sourceMe.R} file should be sourced from.}

\item{weight_names}{The weight names.}

\item{...}{Argumentns pass to \code{get_weights_timeline}.}
}
\value{
A data.table of weight evolution
}
\description{
Get weight evolution data for a simulation replicate.
}
