% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_weight_evo.R
\name{get_weights_prop}
\alias{get_weights_prop}
\title{Get weight data from \code{generation} output.}
\usage{
get_weights_prop(
  gen_data,
  weight_w,
  min_w_val = -1.01,
  max_w_val = 1.01,
  steps = 50
)
}
\arguments{
\item{gen_data}{A list object produced by \code{generation}.}

\item{weight_w}{The weight identity to be retrieved.}

\item{min_w_val}{The lower bound of the scaled weight values.}

\item{max_w_val}{The upper bound.}

\item{steps}{The number of steps between bounds.}
}
\value{
Returns a \code{data.table}
with columns \code{weight_id}, \code{weight_value} (-1 -- +1), and
\code{weight_prop}.
}
\description{
A function to return scaled weight proportion data from weight values
returned by the \code{generation} function. Returns a \code{data.table}
with columns \code{weight_id}, \code{weight_value} (-1 -- +1), and
\code{weight_prop}.
}
